Keep AkkaTypedDocSpec/Test clear of set up code
